+++ server/routes/subscriberRoutes.js
@@ -4,7 +4,7 @@ const { check } = require("express-validator");
 const multer = require('multer');
 const { register, verify, login, forgetpassword, forgetpasswordverify, updatepassword } = require('../controller/subscriberController');
 const { profile, update, emailChange, verify1, passwordChange, profileImageUpdate, profileImageView } = require('../controller/subscriberProfile');
-const { courseHome,payment,verification,myCourses } = require('../controller/subscriberContentController');
+const { myCourses } = require('../controller/subscriberContentController');
 const { isLoggedIn } = require('../middleware/isLoggedInmiddleware');
 const { protect } = require('../middleware/authMiddleware');
 
@@ -104,22 +104,6 @@ router.post(
         check("new_password", "Passwrod should not be empty "),
     ], passwordChange
 );
-
-router.post(
-    '/courseHome',
-    [
-        check("courseId", "course ID is required.")
-    ], courseHome
-)
-
-router.post(
-    '/payment',isLoggedIn,
-    [],payment
-)
-
-router.post(
-    '/verification',
-    [],verification)
 
 router.get(
     '/myCourses',
